Extract widget in separate file
